feat: Theme and Format Enhancments
Added parsing for a icon field in the json instead of having to contruct the name with the icon.

Added parsing for a theme field. If specified will set the theme to be used for the (sub)menu.

For generated added the ability to specify an exec for the dynamic menu option instead of a submen for each

Updated examples